refactor(ui): extract SourceSection / SenseList / SenseBlock from popup
DefinitionPopup.tsx was ~250 LOC defining four React components +
~100 LOC of inline styles. Splitting into:

  - DefinitionPopup.tsx   (top-level, ~85 LOC)
  - SourceSection.tsx     (per-hit renderer, ~45 LOC)
  - senseBlocks.tsx       (SenseList + SenseBlock, ~55 LOC)
  - popupStyles.ts        (shared StyleSheet, ~115 LOC)

Pure mechanical refactor. No behaviour change, no public API change,
no test changes — all 310 existing tests pass without modification.
The popup-test mock for RN primitives still works since we render
host-string elements and the test asserts on tree text, not component
identity.

Sets up issue #6 (source-format metadata to popup) to land cleanly:
the format-mode switch will live inside SourceSection, isolated from
the rest of the popup logic.

Closes #8.
